+++ Assets/Scripts/Interface/Unit/IEndurance.cs
@@ -2,11 +2,11 @@
 
 public interface IEndurance
 {
-    public float MaxEndurance { get;}
+    public int MaxEndurance { get;}
     public float CurrentEndurance { get; }
 
     public void Initialize();
     public void AddEndurance(float value);
     public void RemoveEndurance(float value);
-    public void IncreaseMaxEndurance(float value);
+    public void IncreaseMaxEndurance(int value);
 }
