+++ Assets/Scripts/ScriptableObjects/Gameplay/Unit/SO_UnitEndurance.cs
@@ -4,6 +4,6 @@ namespace ScriptableObjects.Unit
 {
     public abstract class SO_UnitEndurance : ScriptableObject
     {
-        [field: SerializeField] public float MaxEndurance { get; protected set; } = 1;
+        [field: SerializeField] public int MaxEndurance { get; protected set; } = 1;
     }
 }
